backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0
ollama==0.1.6
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
qdrant-client==1.12.1
boto3==1.35.0
stripe==7.0.0

